Changed menu + a lot of other stuff
Changed the menu colours so the background is white and the text is pink. Also added a bottom border in pink to mimic the real Cosmo website.

Added the Cosmopolitan logo

Changed some classnames to be easier to understand, e.i. description to cardDescription (home cards)

Added all the tips so they are in the right place and with an icon as well. Added the links provided and if the text didn't have links/text I added that +CSS.

Made the footer white and added text

Added the "step pictures" to the home page along with the text we used to have (me and rebecka) +CSS.

Changed the accordion text in size and placement as well as the tips part  +CSS (me and Rebecka)

Changed "Home Article" so it has one column and thus easier for the eye to read. Did the same under Fossil Fuel but that article needs more text (have to do the same with the rest of the articles + CSS.