Changed home card pictures
Changed the home card pictures to cropped versions of the new image.

Added the new images as hero pictures.

Changed diagram key colours to more suitable ones.

Made the blue headline colour a darker blue so it's better contrast. Added the same blue to the menu as well (while hovering).

Made the links in the tip section pink and underlined while hovering over them.

Moved the cosmo logo to line up with the hero pictures.

+ additional CSS

Co-Authored-By: rebeckanasstrom <rebeckanasstrom@users.noreply.github.com>